asyncapi/requirements.txt
aiohttp==3.7.4.post0
aioredis==2.0.0
anyio==3.3.3
asgi-lifespan==1.0.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
cryptography==3.4.7
elasticsearch==7.11.0
environs==9.3.1
fastapi==0.63.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.7
httpx==0.20.0
idna==3.1
iniconfig==1.1.1
marshmallow==3.11.1
multidict==5.1.0
orjson==3.5.0
packaging==21.0
pluggy==1.0.0
pre-commit==2.12.0
py==1.10.0
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.5
pytest-aiohttp==0.3.0
pytest-asyncio==0.15.1
python-dotenv==0.17.0
python-json-logger==2.0.1
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
yarl==1.6.3
